+++ src/content/index.js
@@ -4,3 +4,5 @@ export Row from './columns/content-row.component';
 export Masonry from './masonry/content-masonry.component';
 export InputRow from './input-row/content-input-row.component';
 export ReadonlyRow from './input-row/content-readonly-row.component';
+export InputColumn from './input-column/content-input-column.component';
+export ReadonlyColumn from './input-column/content-readonly-column.component';