+++ src/content/input-column/content-readonly-column.component.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReadonlyRow from '../input-row/content-readonly-row.component';
+
+const ContentReadonlyColumn = props => (<ReadonlyRow asColumn {...props} />);
+export default ContentReadonlyColumn;